requirements.txt
streamlit>=1.35.0
google-generativeai>=0.4.0
pypdf>=4.1.0
pandas>=2.2.0
python-dotenv>=1.0.1
